update: 테스트 실행과 제출이 가능한 footer 추가
