--- frontend/src/components/Simulation/index.ts
@@ -1,1 +1,0 @@
-export * from './Simulator';
