--- frontend/src/components/Submission/index.tsx
@@ -1,1 +1,0 @@
-export { Submission } from './Submission';
