+++ frontend/src/pages/CompetitionPage.tsx
@@ -1,9 +1,9 @@
 import { css } from '@style/css';
 
-import { useEffect, useMemo, useState } from 'react';
+import { useContext, useEffect, useMemo, useState } from 'react';
 import { useParams } from 'react-router-dom';
 
-import { HStack, VStack } from '@/components/Common';
+import { Button, HStack, Modal, Space, VStack } from '@/components/Common';
 import CompetitionHeader from '@/components/Competition/CompetitionHeader';
 import CompetitionProblemSelector from '@/components/Competition/CompetitionProblemSelector';
 import { CompetitionProvider } from '@/components/Competition/CompetitionProvider';
@@ -11,17 +11,21 @@ import Editor from '@/components/Editor/Editor';
 import { PageLayout } from '@/components/Layout/PageLayout';
 import { ProblemHeader } from '@/components/Problem/ProblemHeader';
 import ProblemViewer from '@/components/Problem/ProblemViewer';
-import { Simulator } from '@/components/Simulation';
-import { Submission } from '@/components/Submission/Submission';
+import { SimulationExecButton } from '@/components/Simulation/SimulationExecButton';
+import { SimulationInputModal } from '@/components/Simulation/SimulationInputModal';
+import { SimulationResultList } from '@/components/Simulation/SimulationResultList';
+import { SubmissionButton } from '@/components/Submission/SubmissionButton';
+import { SubmissionResult } from '@/components/Submission/SubmissionResult';
 import { useCompetitionProblem } from '@/hooks/problem';
 import { useCompetitionProblemList } from '@/hooks/problem/useCompetitionProblemList';
+import { SimulationInput, useSimulation } from '@/hooks/simulation';
 
 export default function CompetitionPage() {
   const { id } = useParams<{ id: string }>();
   const competitionId: number = id ? parseInt(id, 10) : -1;
   const [currentProblemIndex, setCurrentProblemIndex] = useState(0);
-
   const { problemList } = useCompetitionProblemList(competitionId);
+  const simulation = useSimulation();
 
   const currentProblem = useMemo(() => {
     if (problemList.length > 0) {
@@ -44,6 +48,23 @@ export default function CompetitionPage() {
 
   const problemIds = problemList.map((problem) => problem.id);
 
+  const handleSimulate = () => {
+    simulation.run(code);
+  };
+
+  const handleSimulationCancel = () => {
+    simulation.cancel();
+  };
+  const modal = useContext(Modal.Context);
+
+  const handleSaveSimulationInputs = (simulationInputs: SimulationInput[]) => {
+    simulation.changeInputs(simulationInputs);
+  };
+
+  function handleOpenModal() {
+    modal.open();
+  }
+
   return (
     <PageLayout className={style}>
       <CompetitionProvider competitionId={competitionId}>
@@ -56,19 +77,33 @@ export default function CompetitionPage() {
               onChangeProblemIndex={setCurrentProblemIndex}
             />
           </aside>
-          <VStack className={hfullStyle}>
-            <ProblemViewer className={problemStyle} content={problem.content}></ProblemViewer>
-            <HStack className={solutionStyle}>
-              <Editor code={problem.solutionCode} onChangeCode={handleChangeCode}></Editor>
-              <section>
-                <Simulator code={code}></Simulator>
-              </section>
-              <section>
-                <Submission code={code} problemId={currentProblem?.id}></Submission>
-              </section>
-            </HStack>
-          </VStack>
+          <HStack className={css({ height: '100%' })}>
+            <VStack className={hfullStyle}>
+              <ProblemViewer className={problemStyle} content={problem.content}></ProblemViewer>
+              <HStack className={solutionStyle}>
+                <Editor code={problem.solutionCode} onChangeCode={handleChangeCode}></Editor>
+                <section>
+                  <SimulationResultList resultList={simulation.results}></SimulationResultList>
+                  <SubmissionResult></SubmissionResult>
+                </section>
+              </HStack>
+            </VStack>
+            <VStack as="footer" className={footerStyle}>
+              <Button onClick={handleOpenModal}>테스트 케이스 추가하기</Button>
+              <Space></Space>
+              <SimulationExecButton
+                isRunning={simulation.isRunning}
+                onExec={handleSimulate}
+                onCancel={handleSimulationCancel}
+              />
+              <SubmissionButton code={code} problemId={currentProblem?.id}></SubmissionButton>
+            </VStack>
+          </HStack>
         </div>
+        <SimulationInputModal
+          simulationInputs={simulation.inputs}
+          onSave={handleSaveSimulationInputs}
+        ></SimulationInputModal>
       </CompetitionProvider>
     </PageLayout>
   );
@@ -86,7 +121,7 @@ const competitionStyle = css({
   overflow: 'hidden',
 });
 const hfullStyle = css({
-  height: '100%',
+  height: 'calc(100% - 50px)',
 });
 const problemStyle = css({
   width: '50%',
@@ -108,4 +143,11 @@ const asideStyle = css({
   borderColor: 'border',
   padding: '0.5rem',
   height: '100%',
+});
+
+const footerStyle = css({
+  height: '50px',
+  width: '100%',
+  paddingX: '0.5rem',
+  gap: '0.5rem',
 });
